prune-queue-handler
