+++ .github/workflows/release.yml
@@ -6,9 +6,6 @@ on:
     branches:
       - main
   workflow_dispatch: {}
-concurrency:
-  group: ${{ github.workflow }}
-  cancel-in-progress: false
 jobs:
   release:
     runs-on: ubuntu-latest
