+++ src/backend/deny-list/prune-queue-handler.lambda.ts
@@ -1,10 +1,10 @@
-import * as AWS from 'aws-sdk';
+import { DeleteObjectCommand, S3Client } from '@aws-sdk/client-s3';
+import { SQSEvent } from 'aws-lambda';
 import { ENV_DELETE_OBJECT_DATA_BUCKET_NAME } from './constants';
 import { requireEnv } from '../shared/env.lambda-shared';
 
-const s3 = new AWS.S3();
-
-export async function handler(event: AWSLambda.SQSEvent) {
+export async function handler(event: SQSEvent) {
+  const s3Client = new S3Client({});
   console.log(JSON.stringify({ event }));
 
   const bucket = requireEnv(ENV_DELETE_OBJECT_DATA_BUCKET_NAME);
@@ -14,6 +14,8 @@ export async function handler(event: AWSLambda.SQSEvent) {
   for (const record of records) {
     const objectKey = record.body;
     console.log(`deleting s3://${bucket}/${objectKey}`);
-    await s3.deleteObject({ Bucket: bucket, Key: objectKey }).promise();
+    await s3Client.send(
+      new DeleteObjectCommand({ Bucket: bucket, Key: objectKey })
+    );
   }
 }
